completed login_register pages

+++ base/views.py
@@ -1,5 +1,10 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from . models import Room, Topic
 from . forms import RoomForm
 from django.db.models import Q
@@ -11,6 +16,51 @@ from django.db.models import Q
 # ]
 
 # Create your views here.
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+          user = User.objects.get(username=username)
+        except:
+          messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR Password does not exist.')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page= 'register'
+    form = UserCreationForm()
+    context = {'form': form, 'page': page}
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration.')
+    return render(request, 'base/login_register.html', context)
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -25,10 +75,12 @@ def home(request):
     return render(request, 'base/home.html', context)
 def room(request, pk):
     room = Room.objects.get(id = pk)
-    context = {'room': room}
+    room_messages = room.message_set.all()
+    context = {'room': room, 'room_messages': room_messages}
 
     return render(request, 'base/room.html', context)
 # you can use contains functions
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -39,9 +91,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=int(pk))
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allow here.')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -50,8 +106,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id= int(pk))
+
+    if request.user != room.host:
+        return HttpResponse('You are not allow here.')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
